Core: Amends d05e72d

+++ core/src/main/kotlin/com/github/tmarsteel/ktprolog/RandomVariable.kt
@@ -2,4 +2,4 @@ package com.github.tmarsteel.ktprolog
 
 import com.github.tmarsteel.ktprolog.term.Variable
 
-class RandomVariable(private val counter: Int) : Variable("_G" + counter)
+class RandomVariable(private val counter: Long) : Variable("_G" + counter)
